requirements.txt
aioredis==1.3.1
APScheduler==3.6.3
asgiref==3.2.10
beautifulsoup4==4.9.3
celery==5.2.2
channels==2.4.0
channels-redis==2.4.2
coreapi==2.3.3
daphne==2.5.0
Django==2.2.27
django-apscheduler==0.5.1
django-cors-headers==3.4.0
django-easy-audit==1.3.0
django-filter==2.4.0
django-redis==4.12.1
django-simpleui==2020.9
django-user-agents==0.4.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-writable-nested==0.6.1
drf-yasg==1.17.1
Pillow==9.0.1
PyJWT==1.7.1
redis==3.5.3
requests==2.25.1
psutil==5.9.5
pycryptodome==3.18.0
captcha==0.5.0